Updated compare_result file
