+++ xhydro/optimal_interpolation/functions/utilities.py
@@ -1,5 +1,6 @@
 import numpy as np
 import csv
+import matplotlib.pyplot as plt
 
 """
 Retourne une liste qui contient les valeurs d'un fichier CSV
@@ -49,6 +50,45 @@ def initialize_nan_arrays(dimensions, quantities):
         t[i][:] = np.nan
     return tuple(t)
 
+
+"""
+Trouve l'association d'une section à une station.
+Arguments :
+stations (list): Liste qui contient les stations
+section_id (string) : L'indentificateur de la section
+Retourne :
+(string): Vide si la section est introuvable ou la clé d'association entre la station et une section.
+"""
+def find_station_section(stations, section_id):
+    value = ""
+    section_position = 0
+    section_value = 1
+    for i in range(len(stations)):
+        if section_id == stations[i][section_position]:
+            value = stations[i]
+    return value[section_value]
+
+
+"""
+Code to plot results as per user request
+"""
+def plot_results(kge, kge_l1o, nse, nse_l1o):
+    fig, (ax1, ax2) = plt.subplots(2)
+    ax1.scatter(kge, kge_l1o)
+    ax1.set_xlabel("KGE")
+    ax1.set_ylabel("KGE Leave-one-out OI")
+    ax1.axline((0, 0), (1, 1), linewidth=2)
+    ax1.set_xlim(0.3, 1)
+    ax1.set_ylim(0.3, 1)
+
+    ax2.scatter(nse, nse_l1o)
+    ax2.set_xlabel("NSE")
+    ax2.set_ylabel("NSE Leave-one-out OI")
+    ax2.axline((0, 0), (1, 1), linewidth=2)
+    ax2.set_xlim(0.3, 1)
+    ax2.set_ylim(0.3, 1)
+
+    plt.show()
 
 """
 We will use functools.partial to define functions instead of lambdas as it is
